Make config command cross-platform.
