+++ Library/Homebrew/extend/os/system_config.rb
@@ -1,0 +1,5 @@
+require "system_config"
+
+if OS.mac?
+  require "extend/os/mac/system_config"
+end
